+ переписан код для совместимости с модулями
