+++ math/core/models/solution.py
@@ -1,17 +1,14 @@
 from typing import Optional
 from pydantic import BaseModel
 
-
-class Task(BaseModel):
-    task: str
-
 class Point(BaseModel):
     x: Optional[str] = None
     y: Optional[str] = None
 
 class Solution(BaseModel):
-    name: str
     task: str
+    name: Optional[str] = None
+    formula: Optional[str] = None
     focus1: Optional[Point] = None
     focus2: Optional[Point] = None
     eccenter: Optional[str] = None
@@ -20,8 +17,8 @@ class Solution(BaseModel):
     direct2: Optional[str] = None
     semiaxis_a: Optional[str] = None
     semiaxis_b: Optional[str] = None
-    assymptotle1: Optional[str] = None
-    assymptotle2: Optional[str] = None
+    asymptote1: Optional[str] = None
+    asymptote2: Optional[str] = None
     center: Optional[Point] = None
 
 class Request(BaseModel):
