+++ math/core/models/task.py
@@ -1,0 +1,5 @@
+from typing import Optional
+from pydantic import BaseModel
+
+class Task(BaseModel):
+    task: str
