+++ math/core/models/verdict.py
@@ -1,0 +1,6 @@
+from typing import Optional
+from pydantic import BaseModel
+
+class Verdict(BaseModel):
+    ok: bool
+    msg: Optional[str] = None
