--- math/main_gen.py
@@ -1,8 +1,0 @@
-from automaton import generate
-
-# запрос на генерацию
-
-task = generate()
-
-# даём задание
-
